Using poetry to ensure exact versions of dependencies
The code base is incompatible with the newest versions of `jax`,
`jaxlib` and `neural-tangents` and it only works with very specific
library versions. Using `poetry` makes it easy to fix the exact versions
and make it easy for newcomers to get started by simply running `poetry
install` followed by `poetry shell`.